Add if-statement to cleanup.

+++ src/Support/Authenticator.php
@@ -71,7 +71,10 @@ class Authenticator extends Google2FA
      */
     public function cleanupTokens(): void
     {
-        DB::table('2fa_tokens')->where('expires_at', '<=', date('Y-m-d H:i:s'))->delete();
+        if (true === config('google2fa.store_in_cookie')) {
+            // TODO add try/catch
+            DB::table('2fa_tokens')->where('expires_at', '<=', date('Y-m-d H:i:s'))->delete();
+        }
     }
 
     /**
